@@ -1,6 +1,6 @@
 module github.com/b1naryth1ef/sneaker
 
-go 1.17
+go 1.19
 
 require (
 	github.com/alioygur/gores v1.2.2
@@ -12,8 +12,10 @@ require (
 )
 
 require (
+	github.com/akavel/rsrc v0.10.2 // indirect
 	github.com/cpuguy83/go-md2man/v2 v2.0.0-20190314233015-f79a8a8ca69d // indirect
 	github.com/gorilla/websocket v1.4.2 // indirect
+	github.com/josephspurrier/goversioninfo v1.4.0 // indirect
 	github.com/russross/blackfriday/v2 v2.0.1 // indirect
 	github.com/shurcooL/sanitized_anchor_name v1.0.0 // indirect
 	github.com/spkg/bom v1.0.0 // indirect
